Add files via upload
A demo PyTorch implementation using a generated image (entropyless). The user can edit the codebase to adjust the image size, patch size, and other hyperparameters as needed. The function can also be modified to see the performance difference between different images. Currently, at a 256x256 image, the image can achieve 29.6 dB, which is quite close to the value of 30dB of a good-quality lossy compressed image.